rabbit_peer_discovery_config: introduce a way to get integer values
Unlike arbitrary values, if environment variable that corresponds to the key
is blank, the function will fall back to the default. The assumption here is
that a blank value cannot be parsed to an integer in a sensible way and
thus using the default is the best we can do.

Part of rabbitmq/rabbitmq-peer-discovery-consul#41.